feat: tip 게시판 scrap 기능 추가

+++ src/main/java/org/example/tackit/domain/free_post/repository/MemberJPARepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface MemberJPARepository extends JpaRepository<Member, Integer> {
+public interface MemberJPARepository extends JpaRepository<Member, Long> {
     Member findTopByOrderByIdAsc();
     Optional<Member> findByNickname(String nickname);
